requirements.txt
opencv-python
numpy
pandas
scipy
duckdb
torch
torchvision
pillow
matplotlib
streamlit